+++ src/dda/config/model/tools.py
@@ -1,0 +1,34 @@
+# SPDX-FileCopyrightText: 2025-present Datadog, Inc. <dev@datadoghq.com>
+#
+# SPDX-License-Identifier: MIT
+from __future__ import annotations
+
+from typing import Literal
+
+from msgspec import Struct, field
+
+
+class BazelConfig(Struct, frozen=True, forbid_unknown_fields=True):
+    """
+    /// tab | :octicons-file-code-16: config.toml
+    ```toml
+    [tools.bazel]
+    managed = "auto"
+    ```
+    ///
+    """
+
+    managed: bool | Literal["auto"] = "auto"
+
+
+class ToolsConfig(Struct, frozen=True):
+    """
+    /// tab | :octicons-file-code-16: config.toml
+    ```toml
+    [tools]
+    ...
+    ```
+    ///
+    """
+
+    bazel: BazelConfig = field(default_factory=BazelConfig)